All async methods are omlemented in DAL

+++ DataLayer/UnitOfWorks/Localization.UoW.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using DataLayer.EF;
 using DataLayer.Entities;
 using DataLayer.Interfaces;
@@ -34,24 +36,68 @@ namespace DataLayer.UnitOfWorks
             _context.SaveChanges();
         }
 
-        public void Remove<T>(T entity) where T : Entity
+        public async Task CommitAsync()
         {
-            _context.Set<T>().Remove(entity);
+            await _context.SaveChangesAsync();
         }
 
-        public void RemoveRange<T>(IEnumerable<T> entities) where T : Entity
+        public T Remove<T>(T entity) where T : Entity
         {
-            _context.Set<T>().RemoveRange(entities);
+            if (entity == null)
+                throw new ArgumentNullException(nameof(entity));
+            return _context.Set<T>().Remove(entity);
+        }
+
+        public async Task<T> RemoveAsync<T>(T entity) where T : Entity
+        {
+            var task = new Task<T>(()=>Remove(entity));
+            task.Start();
+            return await task;
+        }
+
+        public IEnumerable<T> RemoveRange<T>(IEnumerable<T> entities) where T : Entity
+        {
+            if (entities == null)
+                throw new ArgumentNullException(nameof(entities));
+
+            return _context.Set<T>().RemoveRange(entities);
+        }
+
+        public async Task<IEnumerable<T>> RemoveRangeAsync<T>(IEnumerable<T> entities) where T : Entity
+        {
+            var task = new Task<IEnumerable<T>>(() => RemoveRange(entities));
+            task.Start();
+            return await task;
         }
 
         public T Add<T>(T entity) where T : Entity
         {
+            if (entity == null)
+                throw new ArgumentNullException(nameof(entity));
             return _context.Set<T>().Add(entity);
+        }
+
+        public async Task<T> AddAsync<T>(T entity) where T : Entity
+        {
+            var task = new Task<T>(()=>Add(entity));
+            task.Start();
+
+            return await task;
         }
 
         public IEnumerable<T> AddRange<T>(IEnumerable<T> entities) where T : Entity
         {
+            if (entities == null)
+                throw new ArgumentNullException(nameof(entities));
             return _context.Set<T>().AddRange(entities);
+        }
+
+        public async Task<IEnumerable<T>> AddRangeAsync<T>(IEnumerable<T> entities) where T : Entity
+        {
+            var task = new Task<IEnumerable<T>>(() => AddRange(entities));
+            task.Start();
+
+            return await task;
         }
         #endregion
         #region ILocalization
@@ -60,14 +106,35 @@ namespace DataLayer.UnitOfWorks
             return _cityRepository ?? (_cityRepository = new CityRepository(_context));
         }
 
+        public async Task<ICityRepository> GetCityRepositoryAsync()
+        {
+            var task = new Task<ICityRepository>(GetCityRepository);
+            task.Start();
+            return await task;
+        }
+
         public ICountryRepository GetCountryRepository()
         {
             return _countryRepository ?? (_countryRepository = new CountryRepository(_context));
         }
 
+        public async Task<ICountryRepository> GetCountryRepositoryAsync()
+        {
+            var task = new Task<ICountryRepository>(GetCountryRepository);
+            task.Start();
+            return await task;
+        }
+
         public ILanguageRepository GetLanguageRepository()
         {
             return _languageRepository ?? (_languageRepository = new LanguageRepository(_context));
+        }
+
+        public async Task<ILanguageRepository> GetLanguageRepositoryAsync()
+        {
+            var task = new Task<ILanguageRepository>(GetLanguageRepository);
+            task.Start();
+            return await task;
         }
         #endregion
     }
